fix: move column/bar auto label style to plot level
