+++ __tests__/unit/plots/column/component/label-auto-spec.ts
@@ -5,10 +5,8 @@ import sales from '../../../../data/sales.json';
 import subsales from '../../../../data/subsales.json';
 import ColumnAutoLabel from '../../../../../src/plots/column/component/label-auto';
 import { IShape, Element, ORIGIN, FIELD_ORIGIN } from '../../../../../src/dependents';
-import { DEFAULT_GLOBAL_THEME } from '../../../../../src/theme/default';
 import { isContrastColorWhite } from '../../../../../src/util/color';
 
-const LABEL_THEME = DEFAULT_GLOBAL_THEME.label;
 const LABEL_FILL = 'rgba(44, 53, 66, 0.65)';
 
 describe('Column Auto Label', () => {
@@ -46,6 +44,7 @@ describe('Column Auto Label', () => {
       },
     });
     plot.render();
+    const LABEL_THEME = plot.getLayer().theme.label;
     const labels = plot.getLayer().getLabels();
     const columnLabel = labels[0];
     const labelShapes = columnLabel.getLabels();
@@ -83,6 +82,7 @@ describe('Column Auto Label', () => {
       },
     });
     plot.render();
+    const LABEL_THEME = plot.getLayer().theme.label;
     const labels = plot.getLayer().getLabels();
     const columnLabel = labels[0];
     const labelShapes = columnLabel.getLabels();
@@ -117,6 +117,7 @@ describe('Column Auto Label', () => {
       },
     });
     plot.render();
+    const LABEL_THEME = plot.getLayer().theme.label;
     const coordinateBBox = plot.getView().coordinateBBox;
     const labels = plot.getLayer().getLabels();
     const columnLabel = labels[0];
@@ -213,6 +214,7 @@ describe('Grouped Column Auto Label', () => {
       },
     });
     plot.render();
+    const LABEL_THEME = plot.getLayer().theme.label;
     const labels = plot.getLayer().getLabels();
     const columnLabel = labels[0];
     const labelShapes = columnLabel.getLabels();
