+++ src/plots/bar/theme.ts
@@ -10,11 +10,25 @@ const BAR_DISABLE_STYLE = (style) => {
   return { opacity: opacity * 0.5 };
 };
 
-registerTheme('bar', {
+export const DEFAULT_BAR_THEME = {
+  label: {
+    darkStyle: {
+      fill: '#2c3542',
+      stroke: '#ffffff',
+      fillOpacity: 0.85,
+    },
+    lightStyle: {
+      fill: '#ffffff',
+      stroke: '#ffffff',
+      fillOpacity: 1,
+    },
+  },
   columnStyle: {
     normal: {},
     active: BAR_ACTIVE_STYLE,
     disable: BAR_DISABLE_STYLE,
     selected: { lineWidth: 1, stroke: 'black' },
   },
-});
+};
+
+registerTheme('bar', DEFAULT_BAR_THEME);
