+++ src/plots/column/theme.ts
@@ -10,11 +10,25 @@ const COLUMN_DISABLE_STYLE = (style) => {
   return { opacity: opacity * 0.5, fillOpacity: opacity * 0.5 };
 };
 
-registerTheme('column', {
+export const DEFAULT_COLUMN_THEME = {
+  label: {
+    darkStyle: {
+      fill: '#2c3542',
+      stroke: '#ffffff',
+      fillOpacity: 0.85,
+    },
+    lightStyle: {
+      fill: '#ffffff',
+      stroke: '#ffffff',
+      fillOpacity: 1,
+    },
+  },
   columnStyle: {
     normal: {},
     active: COLUMN_ACTIVE_STYLE,
     disable: COLUMN_DISABLE_STYLE,
     selected: { lineWidth: 1, stroke: 'black' },
   },
-});
+};
+
+registerTheme('column', DEFAULT_COLUMN_THEME);
