+++ src/plots/grouped-column/theme.ts
@@ -1,0 +1,4 @@
+import { registerTheme } from '../../theme';
+import { DEFAULT_COLUMN_THEME } from '../column/theme';
+
+registerTheme('groupedColumn', DEFAULT_COLUMN_THEME);
