+++ src/plots/stacked-bar/layer.ts
@@ -2,7 +2,9 @@ import { deepMix } from '@antv/util';
 import { registerPlotType } from '../../base/global';
 import { LayerConfig } from '../../base/layer';
 import { ElementOption } from '../../interface/config';
-import BaseBarLayer, { BarViewConfig } from '../bar/layer';
+import BaseBarLayer from '../bar/layer';
+import { BarViewConfig } from '../bar/interface';
+import './theme';
 import './component/label';
 import './component/label-auto';
 import { getGeometryByType } from '../../util/view';
