+++ src/plots/stacked-bar/theme.ts
@@ -1,0 +1,4 @@
+import { registerTheme } from '../../theme';
+import { DEFAULT_BAR_THEME } from '../bar/theme';
+
+registerTheme('stackedBar', DEFAULT_BAR_THEME);
