+++ src/plots/stacked-column/layer.ts
@@ -3,7 +3,9 @@ import { registerPlotType } from '../../base/global';
 import { LayerConfig } from '../../base/layer';
 import ConnectedArea from '../../components/connected-area';
 import { ElementOption } from '../../interface/config';
-import BaseColumnLayer, { ColumnViewConfig } from '../column/layer';
+import BaseColumnLayer from '../column/layer';
+import { ColumnViewConfig } from '../column/interface';
+import './theme';
 import './component/label';
 import './component/label-auto';
 import { getGeometryByType } from '../../util/view';
